requirements.txt
# Requirements for Vietstock RSS Crawler System

# FastAPI and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# CORS and middleware
starlette>=0.27.0

# Web scraping and RSS parsing
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.10

# Database (SQLite for lightweight tracking)
# No additional dependencies needed - uses built-in sqlite3

# Data processing
python-dateutil>=2.8.0

# Environment and configuration
python-dotenv>=0.19.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Job scheduling
apscheduler>=3.10.4

# Timezone handling
pytz>=2023.3
